refactor: RoadmapsPageからRoadmapTemplatesSectionを分離
テンプレート一覧セクション（トグル表示・ステップ展開・テンプレート利用）を
独立コンポーネントとして抽出。RoadmapsPage: 384→286行（98行削減）
